refactor: change name for file project
