--- rss-virtual-keyboard/src/utils/delete-lastCharacter.js
@@ -1,4 +1,0 @@
-export function deleteLastCharacter(string) {
-  if (string.length === 1) return '';
-  return string.slice(0, string.length - 1);
-}
